backend/requirements.txt
# Django核心框架
Django==4.2.7
djangorestframework==3.14.0

# 数据库
mysqlclient==2.2.0
pymongo==4.6.0
djongo==1.3.6

# CORS支持
django-cors-headers==4.3.1

# 认证和权限
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# API文档
drf-spectacular==0.26.5

# 环境变量管理
python-decouple==3.8

# 日期时间处理
python-dateutil==2.8.2

# HTTP请求
requests==2.31.0

# AI服务 - 百度千帆
qianfan==0.3.5

# 数据处理
pandas==2.1.3
numpy==1.24.4

# 图像处理（头像上传）
Pillow==10.1.0

# 缓存
django-redis==5.4.0
redis==5.0.1

# 异步任务
celery==5.3.4

# 日志
structlog==23.2.0

# 测试
pytest==7.4.3
pytest-django==4.7.0

# 开发工具
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# 生产环境
gunicorn==21.2.0
whitenoise==6.6.0

# 安全
cryptography==41.0.7 